requirements.txt
# Web Frameworks
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-neo4j>=0.2.0
langchain-experimental>=0.3.0

# Neo4j
neo4j>=5.17.0

# Vector DB & Embeddings
pinecone>=5.0.0
sentence-transformers>=2.5.0

# LLM
groq>=0.4.0
openai>=1.0.0

# Data & Visualization
plotly>=5.18.0
pyvis>=0.3.2
pandas>=2.2.0
numpy>=1.26.0

# API
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Code Quality
flake8>=7.0.0
black>=24.1.0
